+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## ??? (09/12/2014)
 * Snippets are now colored using CodeMirror mode for currently opened file.
+* When focusing into snippet variables, the area scrolls so the variable is in the view.
 
 ## 0.2.1 (08/12/2014)
 * It is possible to prefill variables to widget, try typing `ng-directive myModule` and use alt-S